added authorization

+++ src/todo/todo.service.ts
@@ -1,7 +1,7 @@
-import { Injectable, Param } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { createTodoDto, UpdateTodoDto } from './dto/task.dto';
+import { createTodoDto } from './dto/task.dto';
 import { Status, TODO } from './todo.entity';
 
 @Injectable()
@@ -11,12 +11,14 @@ export class TodoService {
     return await this.todoModel.find();
   }
 
-  async create(createTodo: createTodoDto): Promise<TODO> {
+  async create(createTodo: createTodoDto, userId: string): Promise<TODO> {
     const { title, description } = createTodo;
     const task = new TODO();
+    task.userId = userId;
     task.title = title;
     task.description = description;
     task.status = Status.NOT_DONE;
+
     return await new this.todoModel(task).save();
   }
 
@@ -24,12 +26,27 @@ export class TodoService {
     return await this.todoModel.findOne({ _id: id });
   }
 
-  async Delete(id: string): Promise<TODO> {
-    return await this.todoModel.findByIdAndDelete(id);
+  async Delete(id: string, userId: string): Promise<TODO> {
+    const item = await this.todoModel.findById(id);
+    if (item.userId === userId) {
+      return await this.todoModel.findByIdAndDelete(id);
+    }
+
+    throw new ForbiddenException('You are not authorized to delete');
   }
 
-  async Update(id: string, updateStatus: Status): Promise<TODO> {
+  async Update(
+    id: string,
+    updateStatus: Status,
+    userId: string,
+  ): Promise<TODO> {
     console.log(updateStatus, 'updated status');
-    return await this.todoModel.findByIdAndUpdate(id, { status: updateStatus });
+    const item = await this.todoModel.findById(id);
+    if (item.userId === userId) {
+      return await this.todoModel.findByIdAndUpdate(id, {
+        status: updateStatus,
+      });
+    }
+    throw new ForbiddenException('You are not authorized to make any changes');
   }
 }
